Cleanup and refactor relauncher
